backend/requirements.txt
mesa==3.0.3
numpy==2.3
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
python-dotenv==1.0.0